move auth-related stuff to its own module
