+++ tumblr_api/Cargo.toml
@@ -14,8 +14,9 @@ exclude = [
 
 [features]
 default = ["client"]
-client = ["api", "npf", "dep:reqwest", "dep:thiserror", "dep:veil", "dep:async-lock", "dep:time", "dep:serde_with", "dep:serde-enum-str"]
+client = ["api", "npf", "auth", "dep:reqwest", "dep:thiserror", "dep:veil", "dep:async-lock", "dep:time", "dep:serde_with", "dep:serde-enum-str"]
 api = ["npf", "dep:time", "dep:thiserror"]
+auth = []   # TODO add deps
 npf = []
 
 [dependencies]
